feat: get workspaces

+++ src/main/resources/liquibase/sql/function/mark_message_as_read.sql
@@ -1,7 +1,7 @@
 --liquibase formatted sql
 
 --changeset michael-bill:mark_message_as_read splitStatements:false
-create or replace function mark_message_as_read(p_message_id integer, p_user_id varchar)
+create or replace function mark_message_as_read(p_message_id bigint, p_user_id varchar)
 returns void as $$
 begin
     update messages
