+++ src/main/resources/liquibase/sql/table/board_columns.sql
@@ -3,7 +3,7 @@
 --changeset michael-bill:board_columns_table
 create table if not exists board_columns
 (
-    id integer primary key,
+    id bigserial primary key,
     workspace_id varchar(128) references workspace (name) not null,
     title text check (char_length(title) >= 1) not null,
     level integer not null,
