+++ src/main/resources/liquibase/sql/table/poll.sql
@@ -3,9 +3,9 @@
 --changeset michael-bill:poll_table
 create table if not exists poll
 (
-    id integer primary key,
+    id bigserial primary key,
     author_id bigint references users (id) not null,
-    chat_id integer references chat (id) not null,
+    chat_id bigint references chat (id) not null,
     title text not null,
     variants text[] not null,
     answers integer[] not null
