+++ src/main/resources/liquibase/sql/table/user_workspace.sql
@@ -5,7 +5,7 @@ create table if not exists user_workspace
 (
     user_id bigint references users (id) not null,
     workspace_id varchar(128) references workspace (name) not null,
-    role_id integer references roles (id) not null,
+    role_id bigint references roles (id) not null,
 
     primary key (user_id, workspace_id)
 );
